refactor: consolidate ruby editor action integration tests
Merge monaco-paste.test.js (1 test) and execute-line.test.js (2 tests)
into ruby-editor-actions.test.js to reduce cold-start overhead.

Refs #240
